refactor: Resolve pageContents circular dependency
No jira.
