+++ src/routes/InventoryComponents/InventoryPageContents.js
@@ -1,0 +1,13 @@
+import HybridInventory from './InventoryComponents/HybridInventory';
+import BifrostPage from './InventoryComponents/BifrostPage';
+
+export const pageContents = {
+  hybridInventory: {
+    key: 'hybridInventory',
+    component: HybridInventory,
+  },
+  bifrost: {
+    key: 'bifrost',
+    component: BifrostPage,
+  },
+};
